+++ packages/react-components/react-aria/library/bundle-size/Default.fixture.js
@@ -1,7 +1,0 @@
-import { useARIAButton } from '@fluentui/react-aria';
-
-console.log(useARIAButton);
-
-export default {
-  name: 'ARIA - Default',
-};