+++ packages/react-components/react-menu-grid-preview/library/src/components/MenuGridItem/MenuGridItem.types.ts
@@ -32,6 +32,7 @@ export type MenuGridItemSlots = {
   secondSubAction?: Slot<MenuGridCellProps>;
 };
 
-export type MenuGridItemProps = ComponentProps<Partial<MenuGridItemSlots>>;
+export type MenuGridItemProps = Omit<ComponentProps<Partial<MenuGridItemSlots>>, 'content'> &
+  Pick<Partial<MenuGridItemSlots>, 'content'>;
 
 export type MenuGridItemState = ComponentState<MenuGridItemSlots>;