+++ packages/react-components/react-tabs/library/src/index.ts
@@ -10,6 +10,7 @@ export {
   useTabIndicatorStyles_unstable,
   useTabStyles_unstable,
   useTab_unstable,
+  useTabBase_unstable,
 } from './Tab';
 export type {
   TabRegisterData,
@@ -22,6 +23,8 @@ export type {
   TabListProps,
   TabListSlots,
   TabListState,
+  TabListBaseProps,
+  TabListBaseState,
 } from './TabList';
 export {
   renderTabList_unstable,
@@ -32,10 +35,6 @@ export {
   useTabListContextValues_unstable,
   useTabListStyles_unstable,
   useTabList_unstable,
+  useTabListA11yBehavior_unstable,
+  useTabListBase_unstable,
 } from './TabList';
-
-// Experimental APIs - will be uncommented in experimental release
-// export type { TabBaseProps, TabBaseState } from './Tab';
-// export { useTabBase_unstable, useTabA11yBehavior_unstable } from './Tab';
-// export type { TabListBaseProps, TabListBaseState } from './TabList';
-// export { useTabListBase_unstable, useTabListA11yBehavior_unstable } from './TabList';