+++ packages/web-components/src/tablist/tablist.base.ts
@@ -4,6 +4,10 @@ import { swapStates, toggleState } from '../utils/element-internals.js';
 import { isFocusableElement } from '../utils/focusable-element.js';
 import type { Tab } from '../tab/tab.js';
 import { isTab } from '../tab/tab.options.js';
+import { swapStates, toggleState } from '../utils/element-internals.js';
+import { isFocusableElement } from '../utils/focusable-element.js';
+import { getDirection } from '../utils/index.js';
+import { waitForConnectedDescendants } from '../utils/request-idle-callback.js';
 import { TablistOrientation } from './tablist.options.js';
 
 /**